third_party/WebKit/LayoutTests/media/video-no-timeupdate-before-playback-expected.txt

Test that no timeupdate event fires during loading.

EVENT(loadstart)
EVENT(loadedmetadata)
EVENT(loadeddata)
EVENT(canplay)
EVENT(canplaythrough)
END OF TEST
